+++ Client/client/client.py
@@ -9,9 +9,9 @@ class Client:
         listening = Event()
         self.server_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
 
-    def connect(self, String: address, String: port):
+    def connect(self, address: str, port: str):
         self.server_socket.connect((address, port))
 
-    def join(self, String: username, String: groupname = "public"):
-        message =
+    def join(self, username: str, groupname: str = "public"):
+        message = username
         self.server_socket.sendall(message)
