+++ Protocol/pyspam/__init__.py
@@ -1,0 +1,2 @@
+from .pyspam import *
+del pyspam
