+++ Protocol/pyspam/pyproject.toml
@@ -1,0 +1,9 @@
+[tool.setuptools.packages.find]
+where = ["./."]
+
+[tool.setuptools.package-data]
+"*" = ["*.so", "*.dll"]
+
+[project]
+name = "pyspam"
+version = "1.0.0"
